# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ERP Metalúrgica PCP/MRP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── SUPPLIERS ──────────────────────────────────────────────
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── WAREHOUSES ─────────────────────────────────────────────
  /warehouses:
    get:
      operationId: listWarehouses
      tags: [warehouses]
      summary: List warehouses
      responses:
        "200":
          description: Warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"
    post:
      operationId: createWarehouse
      tags: [warehouses]
      summary: Create warehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWarehouseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
  /warehouses/{id}:
    patch:
      operationId: updateWarehouse
      tags: [warehouses]
      summary: Update warehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWarehouseBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"

  # ─── SECTORS ────────────────────────────────────────────────
  /sectors:
    get:
      operationId: listSectors
      tags: [sectors]
      summary: List sectors
      responses:
        "200":
          description: Sectors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sector"
    post:
      operationId: createSector
      tags: [sectors]
      summary: Create sector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSectorBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sector"
  /sectors/{id}:
    patch:
      operationId: updateSector
      tags: [sectors]
      summary: Update sector
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSectorBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sector"

  # ─── MACHINES ───────────────────────────────────────────────
  /machines:
    get:
      operationId: listMachines
      tags: [machines]
      summary: List machines
      responses:
        "200":
          description: Machines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Machine"
    post:
      operationId: createMachine
      tags: [machines]
      summary: Create machine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMachineBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
  /machines/{id}:
    patch:
      operationId: updateMachine
      tags: [machines]
      summary: Update machine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMachineBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"

  # ─── OPERATIONS ─────────────────────────────────────────────
  /operations:
    get:
      operationId: listOperations
      tags: [operations]
      summary: List operations
      responses:
        "200":
          description: Operations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Operation"
    post:
      operationId: createOperation
      tags: [operations]
      summary: Create operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOperationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
  /operations/{id}:
    patch:
      operationId: updateOperation
      tags: [operations]
      summary: Update operation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOperationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"

  # ─── STOCK ──────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ── Dashboard ──────────────────────────────────────────────
...
    Customer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        document:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
      required: [id, name, active, createdAt]
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        document:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
      required: [name]
    UpdateCustomerBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        document:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]

    # ── Supplier ───────────────────────────────────────────────
    Supplier:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        document:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
      required: [id, name, active, createdAt]
    CreateSupplierBody:
      type: object
      properties:
        name:
          type: string
        document:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
      required: [name]
    UpdateSupplierBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        document:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]

    # ── Warehouse ──────────────────────────────────────────────
    Warehouse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        active:
          type: boolean
      required: [id, name, type, active]
    CreateWarehouseBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
      required: [name, type]
    UpdateWarehouseBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]

    # ── Sector ─────────────────────────────────────────────────
    Sector:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        active:
          type: boolean
      required: [id, name, active]
    CreateSectorBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
      required: [name]
    UpdateSectorBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]

    # ── Machine ────────────────────────────────────────────────
    Machine:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        type:
          type: ["string", "null"]
        sectorId:
          type: ["string", "null"]
        sectorName:
          type: ["string", "null"]
        capacityPerDay:
          type: ["number", "null"]
        active:
          type: boolean
      required: [id, code, name, active]
    CreateMachineBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        type:
          type: ["string", "null"]
        sectorId:
          type: ["string", "null"]
        capacityPerDay:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required: [code, name]
    UpdateMachineBody:
      type: object
      properties:
        code:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        sectorId:
          type: ["string", "null"]
        capacityPerDay:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]

    # ── Operation ──────────────────────────────────────────────
    Operation:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        operationType:
          type: string
        active:
          type: boolean
      required: [id, code, name, operationType, active]
    CreateOperationBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        operationType:
          type: string
      required: [code, name, operationType]
    UpdateOperationBody:
      type: object
      properties:
        code:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        operationType:
          type: ["string", "null"]
        active:
          type: ["boolean", "null"]

    # ── Stock Balance ──────────────────────────────────────────
...