# lib/api-zod/src/generated/types/createOperationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ERP Metalúrgica PCP/MRP API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOperationBody {
  code: string;
  name: string;
  /** @nullable */
  description?: string | null;
  operationType: string;
}